requirements.txt
browser-use>=0.1.0
python-dotenv>=0.19.0
langchain-openai>=0.0.1
apscheduler>=3.10.1
playwright>=1.41.0 